Fix compat override for TalkNet Aligner (#3993)
* Fix compatibility override for TalkNet Aligner


* Remove extraneous logging import
