+++ nemo/collections/tts/models/talknet.py
@@ -30,7 +30,6 @@ from nemo.core import Exportable
 from nemo.core.classes import ModelPT, PretrainedModelInfo, typecheck
 from nemo.core.neural_types import MelSpectrogramType, NeuralType
 from nemo.core.neural_types.elements import LengthsType, MelSpectrogramType, RegressionValuesType, TokenIndex
-from nemo.utils import logging
 from nemo.utils.decorators import deprecated
 
 
@@ -415,17 +414,3 @@ class TalkNetSpectModel(SpectrogramGenerator, Exportable):
         model.add_module('_pitch_model', TalkNetPitchModel.from_pretrained(model_name, *args, **kwargs))
         model.add_module('_durs_model', TalkNetDursModel.from_pretrained(model_name, *args, **kwargs))
         return model
-
-    def load_state_dict(self, state_dict, strict=True):
-        """Stopgap measure to keep old checkpoints working before full model deprecation.
-
-        This override fiddles with the state_dict in order to keep functionality from old checkpoints after the
-        switch from torch_stft. It will be removed when this model is deprecated.
-        """
-        if 'preprocessor.featurizer.stft.forward_basis' in state_dict:
-            logging.warning("Loading old checkpoint, defaulting to hann window.")
-            window_dim = state_dict['preprocessor.featurizer.stft.forward_basis'].shape[-1]
-            state_dict['preprocessor.featurizer.window'] = torch.hann_window(window_dim)
-            del state_dict['preprocessor.featurizer.stft.forward_basis']
-            del state_dict['preprocessor.featurizer.stft.inverse_basis']
-        super().load_state_dict(state_dict, strict=strict)
